Tree building works
I can now build trees leaf to root without a timeout and still get
every cell on every tree.  The trick is twofold.  First, cells only
send their own discover messages if all physical ports are connected or
if they have received a minimum number of discover messages for other
trees.  Second, the cells are connected from the outside inward.  This
procedure makes it highly likely that all cells that won’t have all
their physical ports connected (because they are on the edge) will be
connected to all their neighbors before receiving the minimum number of
discover messages for other trees.

+++ cellagent/src/app_message.rs
@@ -1,6 +1,6 @@
 use std::{fmt,
           ops::{Deref},
-          sync::atomic::{AtomicUsize, Ordering},
+          sync::atomic::{AtomicU64, Ordering},
 };
 
 use serde_json;
@@ -16,8 +16,8 @@ use crate::utility::{ByteArray, S};
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq, Serialize, Deserialize)]
 pub struct SenderMsgSeqNo(pub u64);
 impl Deref for SenderMsgSeqNo { type Target = u64; fn deref(&self) -> &Self::Target { &self.0 } }
-static MESSAGE_COUNT: AtomicUsize = AtomicUsize::new(0);
-pub fn get_next_count() -> SenderMsgSeqNo { SenderMsgSeqNo(MESSAGE_COUNT.fetch_add(1, Ordering::SeqCst) as u64) }
+static MESSAGE_COUNT: AtomicU64 = AtomicU64::new(0);
+pub fn get_next_count() -> SenderMsgSeqNo { SenderMsgSeqNo(MESSAGE_COUNT.fetch_add(1, Ordering::SeqCst)) }
 
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq, Serialize, Deserialize)]
 pub enum AppMsgType { // Make sure these match the struct names
